- Clip fade functors (aka easinmg functors) move in to main Clip source module.
