+++ Makefile.in
@@ -23,7 +23,6 @@ headers = \
 	src/qtractorAudioSndFile.h \
 	src/qtractorAudioVorbisFile.h \
 	src/qtractorClip.h \
-	src/qtractorClipFadeFunctor.h \
 	src/qtractorClipCommand.h \
 	src/qtractorClipSelect.h \
 	src/qtractorCommand.h \
@@ -155,7 +154,6 @@ sources = \
 	src/qtractorAudioVorbisFile.cpp \
 	src/qtractorClip.cpp \
 	src/qtractorClipCommand.cpp \
-	src/qtractorClipFadeFunctor.cpp \
 	src/qtractorClipSelect.cpp \
 	src/qtractorCommand.cpp \
 	src/qtractorConnect.cpp \
