--- src/qtractorClipFadeFunctor.h
@@ -1,34 +1,0 @@
-// qtractorClipFadeFunctor.h
-//
-/****************************************************************************
-   Copyright (C) 2010, rncbc aka Rui Nuno Capela. All rights reserved.
-
-   Adapted and refactored from Robert Penner's easing equations.
-   Copyright (C) 2001, Robert Penner.
-
-   This program is free software; you can redistribute it and/or
-   modify it under the terms of the GNU General Public License
-   as published by the Free Software Foundation; either version 2
-   of the License, or (at your option) any later version.
-
-   This program is distributed in the hope that it will be useful,
-   but WITHOUT ANY WARRANTY; without even the implied warranty of
-   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-   GNU General Public License for more details.
-
-   You should have received a copy of the GNU General Public License along
-   with this program; if not, write to the Free Software Foundation, Inc.,
-   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-
-*****************************************************************************/
-
-#ifndef __qtractorClipFadeFunctor_h
-#define __qtractorClipFadeFunctor_h
-
-#include "qtractorClip.h"
-
-
-#endif	// __qtractorClipFadeFunctor_h
-
-
-// end of qtractorClipFade.h
